Cost Directional Distance Function model

The cost directional distance function model is computed by solving an input-oriented directional distance function DEA model for the technical efficiency.

The directional distance function DDF, projects firm $\left( {{\mathbf{x}_o,\mathbf{y}_{o}}} \right)$ to the production frontier in the pre-assigned direction $\mathbf{g}= {\left({{\mathbf{g_{x}^-},\mathbf{g^{+}_y}}} \right)\neq\mathbf{0}_{M+N}}$, $\mathbf{g^{-}_{x}}\mathbb{\in R}^M$ and $\mathbf{g^+_{y}}\mathbb{\in R}^N$. Decomposing cost inefficiency requires defining the input-oriented directional distance functions. Departing from the general definition this implies setting the output directional vector to zero; i.e., $\mathbf{g}= \left({{\mathbf{g_{x}^-}},{\mathbf{g_{y}^+}}} \right)=\:$$\left({{\mathbf{g}_{\textbf{x}}^-,\textbf{0}}} \right)$. Therefore the input-oriented directional distance functions define as follows:

\[T{{I}_{DDF(I)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}}, \textbf{g}_{\textbf{x}}^{{-}}, \textbf{0} \right)=\,\max \,\left\{ \beta_I :\left( {{\textbf{x}}_{o}}-{{\beta_I}}\textbf{g}_{\textbf{x}}^{{-}},{{\textbf{y}}_{o}} \right) \in L(\textbf{y}_o),\ {{\beta_I }}\ge 0 \right\}.\]

The linear programs that allows calculating this measure is:

\[\begin{split} & T{{I}_{DDF(I)}}\left( {{\mathbf{x}}_{o}},{{\mathbf{y}}_{o}} \right)=\underset{{{\beta_I}_{{}}},\lambda }{\mathop{\text{max}}}\,\beta_I \,\, \\ & s.t. \quad \sum\limits_{j\in J}^{{}}{{{\lambda }_{j}}{{x}_{jm}}}\le {{x}_{om}}-\beta_I g_{{{x}_{m}}}^{{-}},\,\,\,m=1,...,M \\ & \quad \quad \sum\limits_{j\in J}^{{}}{{{\lambda }_{j}}{{y}_{jn}}}\ge {{y}_{on}}\,\,n=1,...,N \\ & \quad \quad \sum\limits_{j\in J}^{{}}{{{\lambda }_{j}}}=1 \\ & \quad \quad {{\lambda }_{j}}\ge 0,\,\,j\in J. \end{split}\]

Once again, as in the graph case already presented for the Profit Directional Distance Function model, the choice of directional vector corresponds to the researcher. Customarily, to keep consistency with the radial models, the observed amounts of inputs set the direction: $\mathbf{g}= \left({{\mathbf{g_{x}^-}},{\mathbf{g_{y}^+}}} \right)=\:$$\left({{\mathbf{x}_o,\textbf{0}}}\right)$. In this case it can be shown that the directional model nests the Radial Input Oriented Model. Indeed, if $\left({{\mathbf{g_{x}^-},\mathbf{g^+_y}}} \right)= \:$$\left( {{\mathbf{x}_o,\mathbf{0}}} \right)$, then $\beta_I^{*}=1-\theta^*$ (in the Cost Radial model. However, other choices are available, which are included as options in BenchmarkingEconomicEfficiency.jl–see the documentation below accompanying this function.

The notion of Nerlovian cost inefficiency corresponds to the decompostion of economic efficiency based on the input-oriented directional distance function: $C{{I}_{DDF\left( I \right)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}},\mathbf{g_{x}^-},{\tilde{\textbf{w}}} \right)$ = $T{{I}_{DDF\left( I \right)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o},\mathbf{g_{x}^-}} \right)$ + $A{{I}_{DDF\left( I \right)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}},\mathbf{g_{x}^-}, {\tilde{\textbf{w}}} \right)$. This results in the following expression:

\[\underbrace{\frac{\sum\limits_{m=1}^{M}{{{w}_{m}}{{x}_{om}}}-C\left( {{\textbf{y}}_{o}}, \textbf{w} \right)}{\sum\limits_{m=1}^{M}{w_{m}^{{}}g_{om}^{-}}}}_{\text{Norm. Cost Inefficiency}}=\underbrace{\beta_{{O}}^{*}}_{\text{Technical Inefficiency}}+\underbrace{A{{I}_{DDF\left( I \right)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}},\mathbf{g_{x}^-},\tilde{\textbf{w}} \right)}_{\text{Norm. Allocative Inefficiency}} \ge 0,\]

Reference

Chapter 8 in Pastor, J.T., Aparicio, J. and Zofío, J.L. (2022) Benchmarking Economic Efficiency: Technical and Allocative Fundamentals, International Series in Operations Research and Management Science, Vol. 315, Springer, Cham.

Example

In this example we compute the cost efficiency directional distance function measure using the option Gx=:Monetary.

using BenchmarkingEconomicEfficiency

X = [2 2; 1 4; 4 1; 4 3; 5 5; 6 1; 2 5; 1.6 8];

Y = [1; 1; 1; 1; 1; 1; 1; 1];

W = [1 1; 1 1; 1 1; 1 1; 1 1; 1 1; 1 1; 1 1];

costddf = deacostddf(X, Y, W, Gx = :Monetary)

Cost DDF DEA Model 
DMUs = 8; Inputs = 2; Outputs = 1
Orientation = Input; Returns to Scale = VRS
Gx = Monetary
──────────────────────────────
   Cost  Technical  Allocative
──────────────────────────────
1   0.0    0.0        0.0
2   1.0    0.0        1.0
3   1.0    0.0        1.0
4   3.0    2.66667    0.333333
5   6.0    6.0        0.0
6   3.0    0.0        3.0
7   3.0    2.0        1.0
8   5.6    1.2        4.4
──────────────────────────────

Estimated economic, technical and allocative efficiency scores are returned with the efficiency function:

efficiency(costddf, :Economic)

8-element Vector{Float64}:
 0.0
 1.0
 1.0
 3.0
 6.0
 3.0
 3.0
 5.6

efficiency(costddf, :Technical)

8-element Vector{Float64}:
 0.0
 0.0
 0.0
 2.666666666666666
 6.0
 0.0
 2.0
 1.2000000000000002

efficiency(costddf, :Allocative)

8-element Vector{Float64}:
 0.0
 1.0
 1.0
 0.3333333333333339
 0.0
 3.0
 1.0
 4.3999999999999995

deacostddf Function Documentation

BenchmarkingEconomicEfficiency.deacostddf — Function

deacostddf(X, Y, W; Gx)

Compute cost efficiency using directional distance function DEA model for inputs X, outputs Y and price of inputs W.

Direction specification:

The direction Gx can be one of the following symbols.

:Ones: use ones.

:Observed: use observed values.

:Mean: use column means.

:Monetary: use direction so that profit inefficiency is expressed in monetary values.

Alternatively, a vector or matrix with the desired directions can be supplied.

Optional Arguments

rts=:VRS: chooses variable returns to scale. For constant returns to scale choose :CRS.

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.