Profit Efficiency Change

Example

In this example we compute the profit efficiency change measure considering a choice of directional vector that returns profit inefficiency in monetary terms:

using BenchmarkingEconomicEfficiency

X1 = [5 3; 2 4; 4 2; 4 8; 7 9]
Y1 = [7 4; 10 8; 8 10; 5 4; 3 6]
W1 = [2 1; 2 1; 2 1; 2 1; 2 1]
P1 = [3 2; 3 2; 3 2; 3 2; 3 2]

X2 = [14 12; 8 10; 10 8; 16 20; 14 17]
Y2 = [18 10; 36 28; 28 36; 18 14; 12 20]
W2 = [3 4; 3 4; 3 4; 3 4; 3 4]
P2 = [3 5; 3 5; 3 5; 3 5; 3 5]

X = Array{Float64,3}(undef, 5, 2, 2);
X[:, :, 1] = X1;
X[:, :, 2] = X2;

Y = Array{Float64,3}(undef, 5, 2, 2);
Y[:, :, 1] = Y1;
Y[:, :, 2] = Y2;

W = Array{Float64,3}(undef, 5, 2, 2);
W[:, :, 1] = W1;
W[:, :, 2] = W2;

P = Array{Float64,3}(undef, 5, 2, 2);
P[:, :, 1] = P1;
P[:, :, 2] = P2;

profteffchddf = deaprofitchange(X, Y, W, P, Gx = :Monetary, Gy = :Monetary)

Profit Efficiency Change DEA Model 
DMUs = 5; Inputs = 2; Outputs = 2; Time periods = 2
Returns to Scale = VRS
Gx = Monetary; Gy = Monetary
──────────────────────────────────────────────────────────────
   Prof.Change     Tech.Change  Alloc.Change  NF.Base  NF.Comp
──────────────────────────────────────────────────────────────
1       -166.0   -52.0                -114.0      1.0      1.0
2        -18.0     0.0                 -18.0      1.0      1.0
3          4.0     3.10727e-16           4.0      1.0      1.0
4       -175.0  -104.0                 -71.0      1.0      1.0
5       -136.0   -62.0                 -74.0      1.0      1.0
──────────────────────────────────────────────────────────────

Estimated economic, technical and allocative efficiency change scores are returned with the effchange function:

effchange(profteffchddf, :Economic)

5-element Vector{Float64}:
 -166.00000000000114
  -18.000000000001137
    3.999999999998863
 -175.00000000000114
 -136.00000000000114

effchange(profteffchddf, :Technical)

5-element Vector{Float64}:
  -51.99999999999998
    0.0
    3.1072705384326033e-16
 -104.0
  -62.000000000000014

effchange(profteffchddf, :Allocative)

5-element Vector{Float64}:
 -114.00000000000117
  -18.000000000001137
    3.9999999999988627
  -71.00000000000114
  -74.00000000000112

DEA Profit Efficiency Change Functions Documentation

BenchmarkingEconomicEfficiency.deaprofitchangeadd — Function

deaprofitchangeadd(X, Y, W, P, model)

Compute profit efficiency change using data envelopment analysis weighted additive model for inputs X, outputs Y, price of inputs W, and price of outputs P.

Model specification:

:Ones: standard additive DEA model.

:MIP: Measure of Inefficiency Proportions. (Charnes et al., 1987; Cooper et al., 1999)

:Normalized: Normalized weighted additive DEA model. (Lovell and Pastor, 1995)

:RAM: Range Adjusted Measure. (Cooper et al., 1999)

:BAM: Bounded Adjusted Measure. (Cooper et al, 2011)

:Custom: User supplied weights.

Optional Arguments

rhoX: matrix of weights of inputs. Only if model=:Custom.

rhoY: matrix of weights of outputs. Only if model=:Custom.

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.deaprofitchangerussell — Function

deaprofitchangerussell(X, Y, W, P)

Compute profit efficiency change using Russell data envelopment analysis for inputs X, outputs Y, price of inputs W, and price of outputs P.

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.deaprofitchangeerg — Function

deaprofitchangeerg(X, Y, W, P)

Compute profit efficiency change using data envelopment analysis Enhanced Russell Graph Slack Based Measure model for inputs X, outputs Y, price of inputs W, and price of outputs P.

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.deaprofitchange — Function

deaprofitchange(X, Y, W, P; Gx, Gy)

Compute profit efficiency change using data envelopment analysis model for inputs X, outputs Y, price of inputs W, and price of outputs P.

Direction specification:

The directions Gx and Gy can be one of the following symbols.

:Zeros: use zeros.

:Ones: use ones.

:Observed: use observed values.

:Mean: use column means.

:Monetary: use direction so that profit inefficiency is expressed in monetary values.

:Euclidean: use prices normalized by the Euclidean norm of all input and output prices so profit inefficiency is the Euclidean distance between profit hyperplanes.

Alternatively, a vector or matrix with the desired directions can be supplied.

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.deaprofitchangeholder — Function

deaprofichangetholder(X, Y, W, P; l)

Compute profit efficiency change using data envelopment analysis Hölder model for inputs X, outputs Y, price of inputs W, and price of outputs P.

Hölder norm l specification

1.

2.

Inf.

Optional Arguments

weigt=false: set to true for weighted (weakly) Hölder distance function.

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.deaprofitchangemddf — Function

deaprofitchangemddf(X, Y, W, P; Gx, Gy)

Compute profit efficiency change using Modified DDF data envelopment analysis model for inputs X, outputs Y, price of inputs W, and price of outputs P.

Direction specification:

The directions Gx and Gy can be one of the following symbols.

:Observed: use observed values.

Alternatively, a vector or matrix with the desired directions can be supplied.

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.deaprofitchangerddf — Function

deaprofitchangerddf(X, Y, W, P, measure)

Compute profit change efficiency using data envelopment analysis Reverse DDF model for inputs X, outputs Y, price of inputs W, price of outputs P, and efficiency measure.

Measure specification:

:ERG: Enhanced Russell Graph Slack Based Measure.

:MDDF: Modified Directional Distance Function.

Direction specification:

For the Modified Directional Distance Function, the directions Gx and Gy can be one of the following symbols.

:Observed: use observed values.

Alternatively, a vector or matrix with the desired directions can be supplied.

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

atol=1e-6: tolerance for DMU to be considered efficient.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.deaprofitchangegda — Function

deaprofitchangegda(X, Y, W, P, measure)

Compute profit efficiency change using data envelopment analysis General Direct Approach model for inputs X, outputs Y, price of inputs W, price of outputs P, and efficiency measure.

Measure specification:

:ERG: Enhanced Russell Graph (or Slack Based Measure (SBM)).

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

atol=1e-6: tolerance for DMU to be considered efficient.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.effchange — Method

effchange(model::ProfitChangeDEAModel)

Return efficiency change of a profit change DEA model.

Optional Arguments

type=Economic: type of efficiency change scores to return.

Type specification:

:Economic: returns economic efficiency change of the model.

:Technical: returns technical efficiency change.

:Allocative: returns allocative efficiency change.