Profit Hölder model

The profit Hölder model is computed by solving a graph Hölder DEA model for the technical efficiency.

In this example we compute the profit efficiency Hölder L1 measure under variable returns to scale:

julia> X = [2; 4; 8; 12; 6; 14; 14; 9.412];

julia> Y = [1; 5; 8; 9; 3; 7; 9; 2.353];

julia> W = [1; 1; 1; 1; 1; 1; 1; 1];

julia> P = [2; 2; 2; 2; 2; 2; 2; 2];

julia> profitholderl1 = deaprofitholder(X, Y, W, P, l = 1)
Profit Hölder L1 DEA Model 
DMUs = 8; Inputs = 1; Outputs = 1
Returns to Scale = VRS
────────────────────────────────
   Profit  Technical  Allocative
────────────────────────────────
1   4.0          0.0       4.0
2   1.0          0.0       1.0
3   0.0          0.0       0.0
4   1.0          0.0       1.0
5   4.0          3.0       1.0
6   4.0          2.0       2.0
7   2.0          0.0       2.0
8   6.353        6.0       0.353
────────────────────────────────

Estimated economic, technical and allocative efficiency scores are returned with the efficiency function:

julia> efficiency(profitholderl1, :Economic)
8-element Vector{Float64}:
 4.0
 1.0
 0.0
 1.0
 4.0
 4.0
 2.0
 6.353

julia> efficiency(profitholderl1, :Technical)
8-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 2.9999999999999996
 2.000000000000001
 0.0
 6.0

julia> efficiency(profitholderl1, :Allocative)
8-element Vector{Float64}:
 4.0
 1.0
 0.0
 1.0
 1.0000000000000004
 1.9999999999999991
 2.0
 0.35299999999999976

deaprofitholder Function Documentation

BenchmarkingEconomicEfficiency.deaprofitholder — Function

deaprofitholder(X, Y, W, P; l)

Compute profit efficiency using data envelopment analysis Hölder model for inputs X, outputs Y, price of inputs W, and price of outputs P.

Hölder norm l specification

1.

2.

Inf.

Optional Arguments

names: a vector of strings with the names of the decision making units.

Examples

julia> X = [1 1; 1 1; 0.75 1.5; 0.5 2; 0.5 2; 2 2; 2.75 3.5; 1.375 1.75];

julia> Y = [1 11; 5 3; 5 5; 2 9; 4 5; 4 2; 3 3; 4.5 3.5];

julia> P = [2 1; 2 1; 2 1; 2 1; 2 1; 2 1; 2 1; 2 1];

julia> W = [2 1; 2 1; 2 1; 2 1; 2 1; 2 1; 2 1; 2 1];

julia> deaprofitholder(X, Y, W, P, l = 1)
Profit Hölder L1 DEA Model 
DMUs = 8; Inputs = 2; Outputs = 2
Returns to Scale = VRS
───────────────────────────────────
   Profit     Technical  Allocative
───────────────────────────────────
1     1.0   0.0                 1.0
2     1.0  -1.44329e-15         1.0
3     0.0   0.0                 0.0
4     1.0   0.0                 1.0
5     1.0   0.0                 1.0
6     4.0   1.0                 3.0
7     6.0   2.0                 4.0
8     2.0   0.5                 1.5
───────────────────────────────────