Profit Reverse Directional Distance Function model

The profit reverse directional distance function model is computed by solving a graph Reverse DDF model for the technical inefficiency.

The advantage of decomposing profit inefficiency with the $RDDF$ is that it relates existing additive measures of technical inefficiency to the popular directional distance function, $DDF$. The $RDDF$ is capable of transforming any additive measure of graph technical inefficiency, $EM(G)$, such as the Enhanced Russell Graph or the Modified Directional Distance Function, into a single scalar measure corresponding to a standard $DDF$. Therefore, given the set of $J$ firms under study, ${F}_{J}$, and their projections on the frontier, denoted by ${\hat{F}_{J}}$, the $RDDF$ assigns a new $DDF$ score $\beta$ to the original $EM(G)$, compatible with the projections ${\hat{F}_{J}}$.

To calculate the $RDDF\left( EM\left( G \right),{{F}_{J}},{{{\hat{F}}}_{J}} \right)$ for firm $\left( {{\mathbf{x}_o,\mathbf{y}_{o}}} \right)$ we need to determine the directional vector $\mathbf{g}= ({{\mathbf{g_{x}^-},\textbf{g}_\textbf{y}^+}})$ connecting the firm to its projection, $\left( {{{\hat{\textbf{x}}}}_{o}},{{{\hat{\textbf{y}}}}_{o}} \right)\in {{\hat{F}}_{J}}$. Afterwards we calculate the value of the $RDDF$. However, when calculating the new $RDDF$ scores we need to differentiate between firms that are deemed technically efficient under $EM(G)$ and those that are technically inefficient. The measure $EM(G)$ splits the sample of firms ${{F}_{J}}$ into two disjoint subsets: ${{F}_{E}}=\left\{ \left( {{\textbf{x}}_{j}},{{\textbf{y}}_{j}} \right)\in {{F}_{J}}:T{{I}_{E{{M(G)}}}}\left( {{\textbf{x}}_{j}},{{\textbf{y}}_{j}} \right)=0 \right\}$ and ${{F}_{J\sim E}}=\left\{ \left( {{\textbf{x}}_{j}},{{\textbf{y}}_{j}} \right)\in {{F}_{J}}:T{{I}_{E{{M(G)}}}}\left( {{x\textbf{}}_{j}},{{\textbf{y}}_{j}} \right)>0 \right\}$. Then,

If $\left( {{\textbf{x}}_{j}},{{\textbf{y}}_{j}} \right)\in {{F}_{J\sim E}}$, define

\[\left( {{\textbf{g}}_{{{\textbf{x}}_{j}}}},{{\textbf{g}}_{{{\textbf{y}}_{j}}}} \right)= \left( \frac{ {{{\hat{\textbf{x}}}}_{j}}-{{\textbf{x}}_{j}} }{T{{I}_{EM\left( G \right)}}\left( {{\textbf{x}}_{j}},{{\textbf{y}}_{j}} \right)}, \frac{{{{\mathbf{\hat{y}}}}_{j}}-{{\mathbf{y}}_{j}}}{T{{I}_{EM\left( G \right)}}\left( {{\mathbf{x}}_{j}},{{\mathbf{y}}_{j}} \right)} \right), \, and \, \, \beta_{RDDF(G)}^{*}=T{{I}_{E{{M}(G)}}}\left( {{\textbf{x}}_{j}},{{\textbf{y}}_{j}} \right)>0.\]

If $\left( {{\textbf{x}}_{j}},{{\textbf{y}}_{j}} \right)\in {{F}_{E}}$, define

\[\beta _{j}^{*}=T{{I}_{EM\left( I \right)}}\left( {{\textbf{x}}_{j}},{{\textbf{y}}_{j}} \right)=0, \, and \, \, \left( {{\textbf{g}}_{{{\textbf{x}}_{j}}}},{{\textbf{g}}_{{{\textbf{y}}_{j}}}} \right)= \left( {{{\vec{\textbf{k}}}}_{jM}},{{{\vec{\textbf{k}}}}_{jN}} \right)\,\in \mathbb{R}_{++}^{M+N},\]

where ${{\vec{\textbf{k}}}_{jM}}\in \mathbb{R}_{++}^{M}$ and ${{\vec{\textbf{k}}}_{jN}}\in \mathbb{R}_{++}^{N}$ are vectors whose $M$ and $N$ components have the same units of measuremenet that $\left( {{\textbf{x}}_{j}},{{\textbf{y}}_{j}} \right)$–making profit inefficiency units' invariant. For consistency we search for a combination that yields a normalization factor for profit inefficiency whose value is equal to that associated to the underlying efficiency measure, i.e, $\sum\limits_{m=1}^{M}{w_{m}^{{}}k_{om}^{-}}+\sum\limits_{n=1}^{N}{p_{n}^{{}}k_{on}^{+}} = NF_{EM(G)}$. This choice of ${{\textbf{k}}_{j}}$ is numerically relevant because it makes the values of the normalized profit inefficiency based on the $RDDF$ and that of the original $EM(G)$ equal, and therefore their normalized allocative efficiencies can be compared to each other–-their technical inefficiency values being null.

We now present the expression corresponding to the profit inefficiency measure and its decomposition associated with the $RDDF$. From the Profit Directional Distance Function model , we know how to gauge and decompose profit inefficiency through the directional distance function. This results in $\Pi{{I}_{{RDDF\left( EM\left( G \right),{{F}_{J}},{{{\hat{F}}}_{J}} \right)}}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}},\mathbf{g_{x}^-}, \mathbf{g_{y}^+}, \tilde{{\textbf{w}}}, \tilde{{\textbf{p}}} \right)$ = $T{{I}_{{RDDF\left( EM\left( I \right),{{F}_{J}},{{{\hat{F}}}_{J}} \right)}}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o},\mathbf{g_{x}^-}},\mathbf{g_{y}^+} \right)$ + $A{{I}_{RDDF\left( EM\left( G \right),{{F}_{J}},{{{\hat{F}}}_{J}} \right)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}},\mathbf{g_{x}^-},\mathbf{g_{y}^+}, \tilde{{\textbf{w}}}, \tilde{{\textbf{p}}} \right)$; i.e.,

\[\begin{split} & \underbrace{\frac{\Pi \left( \mathbf{w},\mathbf{p} \right)-\left( \sum\limits_{n=1}^{N}{{{p}_{n}}{{y}_{on}}}-\sum\limits_{m=1}^{M}{{{w}_{m}}{{x}_{om}}} \right)}{\sum\limits_{m=1}^{M}{w_{m}^{{}}g_{om}^{-}}+\sum\limits_{n=1}^{N}{p_{n}^{{}}g_{on}^{+}}}}_{\text{Norm}\text{. Profit Inefficiency}}= \\ & \quad =\underbrace{\beta _{RDDF\left( EM\left( G \right),{{F}_{J}},{{{\hat{F}}}_{J}} \right)}^{*}}_{\text{Graph Technical Inefficiency}}+\ \,\underbrace{A{{I}_{RDDF\left( EM\left( G \right),{{F}_{J}},{{{\hat{F}}}_{J}} \right)}}\left( {{\mathbf{x}}_{o}},{{\mathbf{y}}_{o}},\mathbf{g}_{\mathbf{x}}^{\mathbf{-}},\mathbf{g}_{\mathbf{y}}^{+},\mathbf{\tilde{w}},\mathbf{\tilde{p}} \right)}_{\text{Norm}\text{. Allocative Inefficiency}} \ge 0. \end{split} \]

where the efficiency score $\beta _{RDDF\left( EM\left( G \right),{{F}_{J}},{{{\hat{F}}}_{J}} \right)}^{*}$ for technically inefficient firms is obtained by solving the $DDF$ program with the associated directional vectors.

Benchmarking Economic Efficiency with Julia offers the possibility of decomposing profit inefficiency based on the $RDDF$ considering as original $EM(G)$ measure the enhanced Russell graph measure.

Reference

Chapter 12 in Pastor, J.T., Aparicio, J. and Zofío, J.L. (2022) Benchmarking Economic Efficiency: Technical and Allocative Fundamentals, International Series in Operations Research and Management Science, Vol. 315, Springer, Cham.

Example

In this example we compute the profit efficiency Reverse directional distance function measure for the Enhanced Russell Graph associated efficiency measure under variable returns to scale:

using BenchmarkingEconomicEfficiency

X = [2; 4; 8; 12; 6; 14; 14; 9.412];

Y = [1; 5; 8; 9; 3; 7; 9; 2.353];

W = [1; 1; 1; 1; 1; 1; 1; 1];

P = [2; 2; 2; 2; 2; 2; 2; 2];

profitrddf = deaprofitrddf(X, Y, W, P, :ERG)

Profit Reverse DDF DEA Model 
DMUs = 8; Inputs = 1; Outputs = 1
Returns to Scale = VRS
Associated efficiency measure = ERG
──────────────────────────────────
     Profit  Technical  Allocative
──────────────────────────────────
1  4.0        0.0         4.0
2  0.5        0.0         0.5
3  0.0        0.0         0.0
4  0.166667   0.0         0.166667
5  0.8        0.6         0.2
6  0.571429   0.52381     0.047619
7  0.285714   0.142857    0.142857
8  0.949449   0.8         0.149449
──────────────────────────────────

Estimated economic, technical and allocative efficiency scores are returned with the efficiency function:

efficiency(profitrddf, :Economic)

8-element Vector{Float64}:
 4.0
 0.5
 0.0
 0.16666666666666666
 0.8
 0.5714285714285715
 0.2857142857142859
 0.9494489071548664

efficiency(profitrddf, :Technical)

8-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 0.6000000000000002
 0.523809523809524
 0.14285714285714296
 0.8

efficiency(profitrddf, :Allocative)

8-element Vector{Float64}:
 4.0
 0.5
 0.0
 0.16666666666666666
 0.19999999999999984
 0.04761904761904756
 0.14285714285714296
 0.14944890715486636

deaprofitrddf Function Documentation

BenchmarkingEconomicEfficiency.deaprofitrddf — Function

deaprofitrddf(X, Y, W, P, measure)

Compute profit efficiency using data envelopment analysis Reverse DDF model for inputs X, outputs Y, price of inputs W, price of outputs P, and efficiency measure.

Measure specification:

:ERG: Enhanced Russell Graph Slack Based Measure.

:MDDF: Modified Directional Distance Function.

Direction specification:

For the Modified Directional Distance Function, the directions Gx and Gy can be one of the following symbols.

:Observed: use observed values.

Alternatively, a vector or matrix with the desired directions can be supplied.

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

atol=1e-6: tolerance for DMU to be considered efficient.

names: a vector of strings with the names of the decision making units.