Revenue Efficiency Change

Example

In this example we compute the revenue efficiency change measure considering a choice of directional vector that returns revenue inefficiency in monetary terms:

using BenchmarkingEconomicEfficiency

X1 = [5 3; 2 4; 4 2; 4 8; 7 9]
Y1 = [7 4; 10 8; 8 10; 5 4; 3 6]
P1 = [3 2; 3 2; 3 2; 3 2; 3 2]

X2 = [14 12; 8 10; 10 8; 16 20; 14 17]
Y2 = [18 10; 36 28; 28 36; 18 14; 12 20]
P2 = [3 5; 3 5; 3 5; 3 5; 3 5]

X = Array{Float64,3}(undef, 5, 2, 2);
X[:, :, 1] = X1;
X[:, :, 2] = X2;

Y = Array{Float64,3}(undef, 5, 2, 2);
Y[:, :, 1] = Y1;
Y[:, :, 2] = Y2;

P = Array{Float64,3}(undef, 5, 2, 2);
P[:, :, 1] = P1;
P[:, :, 2] = P2;

reveffchddf = dearevenuechangeddf(X, Y, P, Gy = :Monetary)

Revenue Efficiency Change DEA Model 
DMUs = 5; Inputs = 2; Outputs = 2; Time periods = 2
Returns to Scale = VRS
Gy = Monetary
──────────────────────────────────────────────────────────────────
       Rev.Change     Tech.Change   Alloc.Change  NF.Base  NF.Comp
──────────────────────────────────────────────────────────────────
1  -144.0          -134.0          -10.0              1.0      1.0
2    -2.84217e-14     0.0           -2.84217e-14      1.0      1.0
3     0.0             7.54952e-16   -7.54952e-16      1.0      1.0
4  -117.0          -105.5          -11.5              1.0      1.0
5  -103.0          -108.0            5.0              1.0      1.0
──────────────────────────────────────────────────────────────────

Estimated economic, technical and allocative efficiency change scores are returned with the effchange function:

effchange(reveffchddf, :Economic)

5-element Vector{Float64}:
 -144.0
   -2.842170943040401e-14
    0.0
 -117.0
 -103.0

effchange(reveffchddf, :Technical)

5-element Vector{Float64}:
 -134.00000000000003
    0.0
    7.549516567451065e-16
 -105.5
 -108.0

effchange(reveffchddf, :Allocative)

5-element Vector{Float64}:
  -9.999999999999968
  -2.842170943040401e-14
  -7.549516567451065e-16
 -11.5
   5.0

DEA Revenue Efficiency Change Functions Documentation

BenchmarkingEconomicEfficiency.dearevenuechange — Function

dearevenuechange(X, Y, P, model)

Compute revenue efficiency change using data envelopment analysis radial model for inputs X, outputs Y, and price of outputs P.

Optional Arguments

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.dearevenuechangeadd — Function

dearevenuechangeadd(X, Y, P, model)

Compute revenue efficiency change using data envelopment analysis weighted additive model for inputs X, outputs Y, and price of outputs P.

Model specification:

:Ones: standard additive DEA model.

:MIP: Measure of Inefficiency Proportions. (Charnes et al., 1987; Cooper et al., 1999)

:Normalized: Normalized weighted additive DEA model. (Lovell and Pastor, 1995)

:RAM: Range Adjusted Measure. (Cooper et al., 1999)

:BAM: Bounded Adjusted Measure. (Cooper et al, 2011)

:Custom: User supplied weights.

Optional Arguments

rhoY: matrix of weights of outputs. Only if model=:Custom.

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.dearevenuechangerussell — Function

dearevenuechangerussell(X, Y, W, P)

Compute revenue efficiency change using Russell data envelopment analysis for inputs X, outputs Y, and price of outputs P.

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.dearevenuechangeddf — Function

dearevenuechangeddf(X, Y, W, P; Gx, Gy)

Compute revenue efficiency change using data envelopment analysis model for inputs X, outputs Y, and price of outputs P.

Direction specification:

The direction Gy can be one of the following symbols.

:Zeros: use zeros.

:Ones: use ones.

:Observed: use observed values.

:Mean: use column means.

:Monetary: use direction so that revenue inefficiency is expressed in monetary values.

Alternatively, a vector or matrix with the desired directions can be supplied.

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.dearevenuechangeholder — Function

derevenuechangetholder(X, Y, W, P; l)

Compute revenue efficiency change using data envelopment analysis Hölder model for inputs X, outputs Y, and price of outputs P.

Hölder norm l specification

1.

2.

Inf.

Optional Arguments

weigt=false: set to true for weighted (weakly) Hölder distance function.

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.dearevenuechangerddf — Function

dearevenuechangerddf(X, Y, W, P, measure)

Compute revenue change efficiency using data envelopment analysis Reverse DDF model for inputs X, outputs Y, price of outputs P, and efficiency measure.

Measure specification:

:ERG: Enhanced Russell Graph Slack Based Measure.

Direction specification:

For the Modified Directional Distance Function, the directions Gx and Gy can be one of the following symbols.

:Observed: use observed values.

Alternatively, a vector or matrix with the desired directions can be supplied.

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

atol=1e-6: tolerance for DMU to be considered efficient.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.dearevenuechangegda — Function

dearevenuechangegda(X, Y, W, P, measure)

Compute revenue efficiency change using data envelopment analysis General Direct Approach model for inputs X, outputs Y, price of outputs P, and efficiency measure.

Measure specification:

:ERG: Enhanced Russell Graph (or Slack Based Measure (SBM)).

Optional Arguments

monetary=false: decomposition in normalized terms. Monetary terms if true.

atol=1e-6: tolerance for DMU to be considered efficient.

names: a vector of strings with the names of the decision making units.

BenchmarkingEconomicEfficiency.effchange — Method

effchange(model::RevenueChangeDEAModel)

Return efficiency change of a revenue change DEA model.

Optional Arguments

type=Economic: type of efficiency change scores to return.

Type specification:

:Economic: returns economic efficiency change of the model.

:Technical: returns technical efficiency change.

:Allocative: returns allocative efficiency change.