Cost Directional Distance Function model

The cost directional distance function model is computed by solving a directional distance function DEA model for the technical efficiency.

In this example we compute the cost efficiency directional distance function measure under variable returns to scale:

julia> X = [2 2; 1 4; 4 1; 4 3; 5 5; 6 1; 2 5; 1.6 8];

julia> Y = [1; 1; 1; 1; 1; 1; 1; 1];

julia> W = [1 1; 1 1; 1 1; 1 1; 1 1; 1 1; 1 1; 1 1];

julia> costddf = deacostddf(X, Y, W, Gx = :Monetary)
Cost DDF DEA Model 
DMUs = 8; Inputs = 2; Outputs = 1
Orientation = Input; Returns to Scale = VRS
Gx = Monetary
──────────────────────────────
   Cost  Technical  Allocative
──────────────────────────────
1   0.0    0.0        0.0
2   1.0    0.0        1.0
3   1.0    0.0        1.0
4   3.0    2.66667    0.333333
5   6.0    6.0        0.0
6   3.0    0.0        3.0
7   3.0    2.0        1.0
8   5.6    1.2        4.4
──────────────────────────────

Estimated economic, technical and allocative efficiency scores are returned with the efficiency function:

julia> efficiency(costddf, :Economic)
8-element Vector{Float64}:
 0.0
 1.0
 1.0
 3.0
 6.0
 3.0
 3.0
 5.6

julia> efficiency(costddf, :Technical)
8-element Vector{Float64}:
 0.0
 0.0
 0.0
 2.666666666666666
 6.0
 0.0
 2.0
 1.2000000000000002

julia> efficiency(costddf, :Allocative)
8-element Vector{Float64}:
 0.0
 1.0
 1.0
 0.3333333333333339
 0.0
 3.0
 1.0
 4.3999999999999995

deacostddf Function Documentation

BenchmarkingEconomicEfficiency.deacostddf — Function

deacostddf(X, Y, W; Gx)

Compute cost efficiency using directional distance function DEA model for inputs X, outputs Y and price of inputs W.

Direction specification:

The direction Gx can be one of the following symbols.

:Ones: use ones.

:Observed: use observed values.

:Mean: use column means.

:Monetary: use direction so that profit inefficiency is expressed in monetary values.

Alternatively, a vector or matrix with the desired directions can be supplied.

Optional Arguments

rts=:VRS: chooses variable returns to scale. For constant returns to scale choose :CRS.

Examples

julia> X = [2 2; 1 4; 4 1; 4 3; 5 5; 6 1; 2 5; 1.6 8];

julia> Y = [1; 1; 1; 1; 1; 1; 1; 1];

julia> W = [1 1; 1 1; 1 1; 1 1; 1 1; 1 1; 1 1; 1 1];

julia> deacostddf(X, Y, W, Gx = :Monetary)
Cost DDF DEA Model 
DMUs = 8; Inputs = 2; Outputs = 1
Orientation = Input; Returns to Scale = VRS
Gx = Monetary
──────────────────────────────
   Cost  Technical  Allocative
──────────────────────────────
1   0.0    0.0        0.0
2   1.0    0.0        1.0
3   1.0    0.0        1.0
4   3.0    2.66667    0.333333
5   6.0    6.0        0.0
6   3.0    0.0        3.0
7   3.0    2.0        1.0
8   5.6    1.2        4.4
──────────────────────────────