Cost Russell model

The cost Russell model is computed by solving an input-oriented Russell DEA model for the technical efficiency.

Taking the Russell measure as reference, Pastor, Aparicio and Zofío (2022, Ch. 5) present the Fenchel-Mahler inequality obtained from the dual correspondence between the cost function and this measure of input technical inefficiency. The Russell input-oriented measure quantifying the technical inefficiency of a firm can be calculated through DEA methods by solving the following program:

\[\begin{split} & T{{E}_{RM(I)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}} \right)= \underset{\begin{smallmatrix} \pmb{\theta} , \pmb{\lambda} \end{smallmatrix}}{\mathop{\min }}\, \frac{1}{M} \sum\limits_{m=1}^{M}{{{\theta }_{m}}} \\ & s.t. \quad \sum\limits_{j=1}^{J}{{{\lambda }_{j}}{{x}_{jm}}}={{\theta }_{m}}{{x}_{om}}, \quad m=1,...,M \\ & \quad \quad \sum\limits_{j=1}^{J}{{{\lambda }_{j}}{{y}_{jn}}}={{y}_{on}}, \quad n=1,...,N \\ & \quad \quad \sum\limits_{j=1}^{J}{{{\lambda }_{j}}}=1, \\ & \quad \quad {{\theta }_{m}}\le 1, \quad m=1,...,M \\ & \quad \quad {{\lambda }_{j}}\ge 0, \quad j=1,...,J \\ \end{split}\]

In this program, $\theta^*_{m}$ evaluates the relative proportional reduction of input $m, m=1,...,M$. The objective function averages these proportional rates of input contraction. Contrary to the Russell graph DEA model, this program is linear and therefore easy to calculate through the simplex method.

Considering $T{{E}_{RM(I)}}({{\textbf{x}}_{o}},{{\textbf{y}}_{o}})$ we can define is technical inefficiency counterparts as $T{I}_{RM(I)}( {\textbf{x}_{o}},{\textbf{y}_{o}})=1-T{{E}_{RM(I)}} ({\textbf{x}_{o}},{\textbf{y}_{o}})$. It is then possible to decompose cost inefficiency into technical and allocative components: $C{{I}_{RM\left( I \right)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}},{\tilde{\textbf{w}}} \right)$ = $T{{I}_{RM\left( I \right)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}} \right)$ + $A{{I}_{RM\left( I \right)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}}, ,\tilde{\textbf{w}} \right)$. That is

\[\underbrace{\frac{\sum\limits_{m=1}^{M}{{{w}_{m}}{{x}_{om}}} - C \left( \mathbf{y}_o,\mathbf{w} \right)}{ M \min \left\{ {{w}_{1}}{{x}_{o1}},...,{{w}_{M}}{{x}_{oM}}\right\}}}_{\text{Norm. Cost Inefficiency}} = \underbrace{1-\frac{1}{M}\sum\limits_{m=1}^{M}{\theta _{m}^{*}} }_{\text{Input Technical Inefficiency}}+\underbrace{A{{I}_{RM\left( I \right)}}\left( {{\textbf{x}}_{o}},{{\textbf{y}}_{o}},\tilde{\textbf{w}} \right)}_{\text{Norm. Allocative Inefficiency}}\ge 0. \\ \]

Reference

Chapter 5 in Pastor, J.T., Aparicio, J. and Zofío, J.L. (2022) Benchmarking Economic Efficiency: Technical and Allocative Fundamentals, International Series in Operations Research and Management Science, Vol. 315, Springer, Cham.

Example

In this example we compute the cost efficiency Russell measure under variable returns to scale:

using BenchmarkingEconomicEfficiency

X = [2 2; 1 4; 4 1; 4 3; 5 5; 6 1; 2 5; 1.6 8];

Y = [1; 1; 1; 1; 1; 1; 1; 1];

W = [1 1; 1 1; 1 1; 1 1; 1 1; 1 1; 1 1; 1 1];

costrussell = deacostrussell(X, Y, W)

Russell Cost DEA Model 
DMUs = 8; Inputs = 2; Outputs = 1
Orientation = Input; Returns to Scale = VRS
──────────────────────────────
   Cost  Technical  Allocative
──────────────────────────────
1  0.0    0.0        0.0
2  0.5    0.0        0.5
3  0.5    0.0        0.5
4  0.5    0.416667   0.0833333
5  0.6    0.6        0.0
6  1.5    0.166667   1.33333
7  0.75   0.35       0.4
8  1.75   0.4375     1.3125
──────────────────────────────

Estimated economic, technical and allocative efficiency scores are returned with the efficiency function:

efficiency(costrussell, :Economic)

8-element Vector{Float64}:
 0.0
 0.5
 0.5
 0.5
 0.6
 1.5
 0.75
 1.7499999999999998

efficiency(costrussell, :Technical)

8-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.41666666666666663
 0.6
 0.16666666666666663
 0.35
 0.4375

efficiency(costrussell, :Allocative)

8-element Vector{Float64}:
 0.0
 0.5
 0.5
 0.08333333333333337
 0.0
 1.3333333333333335
 0.4
 1.3124999999999998

deacostrussell Function Documentation

BenchmarkingEconomicEfficiency.deacostrussell — Function

deacostrussell(X, Y, W)

Compute cost efficiency using Russell data envelopment analysis for inputs X, outputs Y and price of inputs W.

Optional Arguments

rts=:VRS: chooses variable returns to scale. For constant returns to scale choose :CRS.

monetary=false: decomposition in normalized terms. Monetary terms if true.

names: a vector of strings with the names of the decision making units.